024 apresentacoes fracionadas

+++ src/Amani.ImportadosERP.Application/DTOs/Dashboards/DashboardEstoqueValorizadoDto.cs
@@ -2,10 +2,10 @@ namespace Amani.ImportadosERP.Application.DTOs.Dashboards;
 
 public sealed class DashboardEstoqueValorizadoDto
 {
-    public int QuantidadeTotal { get; set; }
+    public decimal QuantidadeTotal { get; set; }
     public decimal ValorAoCusto { get; set; }
     public decimal ValorAoPrecoVenda { get; set; }
     public decimal LucroPotencialCalculavel { get; set; }
-    public int QuantidadeSemCusto { get; set; }
+    public decimal QuantidadeSemCusto { get; set; }
     public decimal ValorVendaSemCusto { get; set; }
 }
