+++ src/Amani.ImportadosERP.Application/DTOs/Estoque/EstoqueProdutoSaldoDto.cs
@@ -7,5 +7,5 @@ public sealed class EstoqueProdutoSaldoDto
     public Guid ProdutoId { get; set; }
     public string NomeProduto { get; set; } = string.Empty;
     public Guid CategoriaId { get; set; }
-    public int Saldo { get; set; }
+    public decimal Saldo { get; set; }
 }
